+++ frontend/app/layout/HomePageLayout/components/PokemonInstructions/PokemonInstructions.tsx
