+++ frontend/app/layout/HomePageLayout/components/PokemonInstructions/index.ts
